+++ lib/hairballs/ext/kernel_vputs.rb
@@ -3,6 +3,6 @@ module Kernel
   # Does a puts only if IRB is in verbose mode.
   def vputs(*messages)
     messages.map! { |m| "[Hairballs] #{m}" }
-    puts(*messages) if defined?(::IRB) && ::IRB.conf[:VERBOSE]
+    $stdout.puts(*messages) if defined?(::IRB) && ::IRB.conf[:VERBOSE]
   end
 end