+++ lib/hairballs/plugins/awesome_print.rb
@@ -1,7 +1,7 @@
 require 'hairballs'
 
 Hairballs.add_plugin(:awesome_print) do |plugin|
-  plugin.libraries %w(awesome_print)
+  plugin.libraries %w[awesome_print]
 
   plugin.on_load do
     AwesomePrint.irb!