+++ lib/hairballs/plugins/wirble.rb
@@ -2,7 +2,7 @@ require 'hairballs'
 
 # Just loads Wirble.
 Hairballs.add_plugin(:wirble) do |plugin|
-  plugin.libraries %w(wirble)
+  plugin.libraries %w[wirble]
 
   plugin.on_load do
     Wirble.init