+++ lib/hairballs/theme.rb
@@ -12,18 +12,20 @@ class Hairballs
   class Theme
     include Hairballs::LibraryHelpers
 
-    # Just an identifier for the Theme.  Don't name two themes the same
-    # name--that will cause problems.
+    # @!attribute name [rw]
+    #   Just an identifier for the Theme.  Don't name two themes the same
+    #   name--that will cause problems.
     #
-    # @param [Symbol]
+    #   @return [Symbol]
     attr_accessor :name
 
-    # Tells Hairballs to do some hackery to let Themes use gems that aren't
-    # specified in your app's Gemfile.  This alleviates you from having to
-    # declare gems in your Gemfile simply for the sake of managing your personal
-    # IRB preferences.
+    # @!attribute extend_bundler [rw]
+    #   Tells Hairballs to do some hackery to let Themes use gems that aren't
+    #   specified in your app's Gemfile.  This alleviates you from having to
+    #   declare gems in your Gemfile simply for the sake of managing your
+    #   personal IRB preferences.
     #
-    # @param [Boolean]
+    #   @return [Boolean]
     attr_accessor :extend_bundler
 
     # @param name [Symbol]
@@ -36,8 +38,10 @@ class Hairballs
     # Tell IRB to use this Theme.
     def use!
       do_bundler_extending if @extend_bundler
+      vputs "[th:#{@name}] Requiring libs..."
       require_libraries
       set_up_prompt
+      vputs "[th:#{@name}] Done setting up."
     end
 
     # The name of the Theme, but in the format that IRB.conf[:PROMPT] likes (an
@@ -79,7 +83,8 @@ class Hairballs
     end
 
     def set_up_irb_prompt
-      IRB.conf[:PROMPT][irb_name] = @prompt.irb_configuration
+      vputs "[th:#{@name}] Prompt options: #{@prompt.irb_prompt_options}"
+      IRB.conf[:PROMPT][irb_name] = @prompt.irb_prompt_options
       IRB.conf[:PROMPT_MODE] = irb_name
       IRB.CurrentContext.prompt_mode = irb_name if IRB.CurrentContext
     end
@@ -92,7 +97,9 @@ class Hairballs
       elsif @prompt.normal
         ::Pry.config.prompt = -> { @prompt.normal }
       else
-        vputs 'Neither "normal" nor "continued_statement" prompts configured.'
+        # rubocop:disable Metrics/LineLength
+        vputs %([th:#{@name}] Neither "normal" nor "continued_statement" prompts configured.)
+        # rubocop:enable
       end
     end
 
@@ -102,7 +109,7 @@ class Hairballs
           output.printf @prompt.return_format, value.inspect
         end
       else
-        vputs '"return_format" not configured.'
+        vputs %([th:#{@name}] "return_format" not configured.)
       end
     end
   end