+++ History.md
@@ -1,5 +1,20 @@
 # Change Log
 
+## [unreleased] / yyyy-mm-dd
+
+### Improvements
+
+* For `Plugin`s, only try to require libraries if some are given for the plugin.
+* Changed `quick_benchmark` default to 20 iterations (from 100).
+* Changed `Hairballs::LibraryHelpers` to use Fibers instead of Threads for
+  installing dependecies. Threaded installing was proving to be problematic.
+
+### Bug Fixes
+
+* Removed `Prompt#auto_indent`; it never worked. To do auto-indenting, just use
+  `IRB.conf[:AUTO_INDENT]` directly.
+* Fixed incorrect uses of Yard.
+
 ## 0.1.3 / 2015-07-13
 
 ### Improvements